style: extract editor event  and create option type
